mandelbrot Fortran code: fixes warnings with recent versions of gcc (2nd try...)

+++ src/mandel.f90
@@ -9,7 +9,7 @@ contains
 
 subroutine mandelbrot( orig, dx, dy, R, NX, NY, NMAX )
   complex,intent(in) :: orig, dx, dy
-  integer,intent(in) :: NX, NY, NMAX
+  integer*2,intent(in) :: NX, NY, NMAX
   integer*2,intent(inout) :: R(NX,NY)
   complex(kind=kind(0d0)) :: point, z, zn
   integer*2 :: i,j,k
